refactor: configure eslint
