+++ server/package.json
@@ -19,17 +19,17 @@
         "@types/cors": "^2.8.17",
         "@types/express": "^4.17.21",
         "@types/node": "^20.10.4",
+        "@typescript-eslint/eslint-plugin": "^6.14.0",
+        "@typescript-eslint/parser": "^6.14.0",
         "eslint": "^8.55.0",
-        "eslint-config-standard-with-typescript": "^42.0.0",
+        "eslint-config-airbnb-base": "^15.0.0",
+        "eslint-config-prettier": "^9.1.0",
         "eslint-import-resolver-typescript": "^3.6.1",
-        "eslint-plugin-import": "^2.29.0",
+        "eslint-plugin-import": "^2.29.1",
         "eslint-plugin-import-helpers": "^1.3.1",
-        "eslint-plugin-jest": "^27.6.0",
-        "eslint-plugin-n": "^16.4.0",
         "eslint-plugin-prettier": "^5.0.1",
-        "eslint-plugin-promise": "^6.1.1",
-        "eslint-plugin-security": "^1.7.1",
         "jest": "^29.7.0",
+        "prettier": "^3.1.1",
         "prisma": "^5.7.0",
         "ts-node-dev": "^2.0.0",
         "typescript": "^5.3.3"
