+++ server/src/modules/students/model/Student.interface.ts
