+++ server/src/modules/students/use-cases/create-student/create-student-use-case.ts
@@ -1,8 +1,8 @@
-import { type ICreateStudentDTO } from '../../../../adapters/student/create-student-dto.interface'
-import { type IStudentsRepository } from '../../repositories/students-repository.interface'
-import { type UseCase } from '../use-case.interface'
+import { ICreateStudentDTO } from '../../../../adapters/student/create-student-dto.interface'
+import { IStudentsRepository } from '../../repositories/students-repository.interface'
+import { IUseCase } from '../use-case.interface'
 
-class CreateStudentUseCase implements UseCase<ICreateStudentDTO, void> {
+class CreateStudentUseCase implements IUseCase<ICreateStudentDTO, void> {
     constructor(private readonly studentsRepository: IStudentsRepository) {}
 
     async execute({ name, email, ra, cpf }: ICreateStudentDTO): Promise<void> {
