+++ server/src/modules/students/use-cases/create-student/index.ts
@@ -1,4 +1,5 @@
 import { StudentsRepository } from '../../repositories/students-repository'
+
 import { CreateStudentController } from './create-student-controller'
 import { CreateStudentUseCase } from './create-student-use-case'
