+++ server/src/modules/students/use-cases/delete-student/delete-student-use-case.ts
@@ -1,7 +1,7 @@
-import { type IStudentsRepository } from '../../repositories/students-repository.interface'
-import { type UseCase } from '../use-case.interface'
+import { IStudentsRepository } from '../../repositories/students-repository.interface'
+import { IUseCase } from '../use-case.interface'
 
-class DeleteStudentUseCase implements UseCase<string, void> {
+class DeleteStudentUseCase implements IUseCase<string, void> {
     constructor(private readonly studentsRepository: IStudentsRepository) {}
 
     async execute(id: string): Promise<void> {
