+++ server/src/modules/students/use-cases/delete-student/index.ts
@@ -1,4 +1,5 @@
 import { StudentsRepository } from '../../repositories/students-repository'
+
 import { DeleteStudentController } from './delete-student-controller'
 import { DeleteStudentUseCase } from './delete-student-use-case'
