+++ server/src/modules/students/use-cases/get-student/get-student-use-case.ts
@@ -1,8 +1,8 @@
-import { type IStudent } from '../../model/Student'
-import { type IStudentsRepository } from '../../repositories/students-repository.interface'
-import { type UseCase } from '../use-case.interface'
+import { IStudent } from '../../model/Student.interface'
+import { IStudentsRepository } from '../../repositories/students-repository.interface'
+import { IUseCase } from '../use-case.interface'
 
-class GetStudentUseCase implements UseCase<string, IStudent> {
+class GetStudentUseCase implements IUseCase<string, IStudent> {
     constructor(private readonly studentsRepository: IStudentsRepository) {}
 
     async execute(id: string): Promise<IStudent> {
