+++ server/src/modules/students/use-cases/get-student/index.ts
@@ -1,4 +1,5 @@
 import { StudentsRepository } from '../../repositories/students-repository'
+
 import { GetStudentController } from './get-student-controller'
 import { GetStudentUseCase } from './get-student-use-case'
