+++ server/src/modules/students/use-cases/list-students/index.ts
@@ -1,4 +1,5 @@
 import { StudentsRepository } from '../../repositories/students-repository'
+
 import { ListStudentsController } from './list-students-controller'
 import { ListStudentsUseCase } from './list-students-use-case'
