+++ server/src/modules/students/use-cases/update-student/index.ts
@@ -1,4 +1,5 @@
 import { StudentsRepository } from '../../repositories/students-repository'
+
 import { UpdateStudentController } from './update-student-controller'
 import { UpdateStudentUseCase } from './update-student-use-case'
